+++ src/lib.rs
@@ -98,6 +98,7 @@ pub use self::{
     ffi::{module_to_object, object_to_shared_lib, LLVMCompileError, OptLevel},
 };
 
+mod arch;
 pub(crate) mod block_ext;
 pub mod cache;
 mod compiler;
